# swagger/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "CryptoPill",
		"contact": {},
		"version": "1.0",
		"description": "API Docs"
	},
	"servers": [
		{
			"url": "http://180.149.241.208:3048",
			"variables": {}
		},
		{
			"url": "http://localhost:3044",
			"variables": {}
		}
	],
	"paths": {
...
		"/laboratory/updateLab": {
			"post": {
				"tags": ["Laboratory"],
				"summary": "updateSetupProfile",
				"operationId": "updateSetupProfile",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/updateSetupProfileRequest"
							},
							"example": {
								"laboratoryDetails": {
									"workplaces": [
										{
											"address": {
												"address_id": 5,
												"address": "Rama Tower, new Gopal Garden, Chembur - East , Mumbai, 400089",
												"location": {
													"latitude": 46.1672,
													"longitude": -46.353724
												},
												"locality": "sector 11",
												"city": "Mumbai",
												"pincode": "400341"
											},
											"workplace": {
												"workplace_id": 1,
												"workplace_name": "Jara Plutonics Laboratory",
												"is_franchise": 1,
												"franchise_name": "Jara Industries",
												"phone_number": "9898167634",
												"license_number": "AC129ETBMQ56U",
												"gst_number": "12845NKYTLDS0",
												"delivery_customer": 1,
												"delivery_distance": 10
											},
											"employees": [
												{
													"first_name": "lab_assistant4",
													"middle_name": null,
													"last_name": "lab_lastname4",
													"email": "testlab4@gmail.com",
													"gender": "male",
													"birth_date": "1995-11-07",
													"contact_number": "+919066660652",
													"lab_or_pharma_employement_number": "1102",
													"document_type": "ADHAAR",
													"document_number": "12FG1222040",
													"edit_profile": 1
												}
											]
										}
									],
									"user_id": 8
								},
								"user_role": 4
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false
			}
		},
...
		"/doctor/dashboard": {
			"get": {
				"tags": ["Doctor"],
				"summary": "Dashboard",
				"operationId": "Dashboard",
				"parameters": [
					{
						"name": "date",
						"in": "query",
						"description": "",
						"required": true,
						"style": "form",
						"explode": true,
						"schema": {
							"type": "string",
							"example": "2020-10-15"
						}
					},
					{
						"name": "day",
						"in": "query",
						"description": "",
						"required": true,
						"style": "form",
						"explode": true,
						"schema": {
							"type": "string",
							"example": "Monday"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Dashboard"
								},
								"example": {
									"success": true,
									"data": [
										{
											"workplace_id": 1,
											"workplace_name": "Sunny11 Hospptal",
											"time_per_appointment": 20,
											"slots": [
												{
													"id": 1,
													"start_time": "10:00:00",
													"end_time": "12:00:00",
													"blocked_slots": []
												}
											],
											"patient_details": []
										},
										{
											"workplace_id": 2,
											"workplace_name": "AIIMS",
											"time_per_appointment": 30,
											"slots": [
												{
													"id": 3,
													"start_time": "15:00:00",
													"end_time": "17:00:00",
													"blocked_slots": []
												}
											],
											"patient_details": []
										}
									],
									"message": "Success"
								}
							}
						}
					}
				},
				"deprecated": false
			}
		},
...
		"/doctor/updateDelegates": {
			"post": {
				"tags": ["Doctor"],
				"summary": "updateDelegates",
				"operationId": "updateDelegates",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/updateDelegatesRequest"
							},
							"example": {
								"staff_id": 5,
								"workplaces_id": 9,
								"manageAppoinment": 1,
								"blockUnblockSchedue": 0,
								"changeSchedule": 0
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false
			}
		},
		"/doctor/signup/addSetupProfileDetails": {
			"post": {
				"tags": ["Doctor"],
				"summary": "addSetupProfile",
				"operationId": "PostaddSetupProfile1",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/addSetupProfileRequest2"
							},
							"example": {
								"professionalInformation": {
									"drInfo": {
										"birth_date": "1985-08-01",
										"gender": "male",
										"email": "dummydoc@gmail.com",
										"user": {
											"id": 3
										},
										"experience": 6
									},
									"specialites": [1]
								},
								"educationalQualification": {
									"mrdData": {
										"registration_number": "MRC200510084HRAC",
										"registration_year": 2005,
										"registration_council": 2
									},
									"educationDetails": [
										{
											"qualification_id": 1,
											"university_id": 1,
											"year": 2004
										}
									]
								},
								"workplaceInformation": {
									"workplaces": [
										{
											"address": {
												"address": "Sunny11 Tower, new Ambedkar Garden, Chembur - East , Mumbai, 400089",
												"location": {
													"latitude": 46.1052,
													"longitude": -46.383724
												},
												"locality": "sector 12",
												"city": "Mumbai",
												"pincode": "400341"
											},
											"workplace": {
												"workplace_name": "Sunny11 Hospptal",
												"time_per_appointment": 20,
												"consultation_fee": 200
											},
											"schedule": [
												{
													"day": "Monday",
													"start_time": "15:00",
													"end_time": "17:00"
												}
											]
										}
									]
								},
								"otherInformation": {
									"conventionDetails": {
										"medical_convention": "Allopathy",
										"prescription_limit": 14,
										"prescription_days_week_month": "Days"
									},
									"documentDetails": {
										"type": "ADHAAR",
										"number": "12FG12345678"
									},
									"features": [1, 2, 3]
								},
								"user_role": 2
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
		"/doctor/workplaces": {
			"get": {
				"tags": ["Doctor"],
				"summary": "workplaces",
				"operationId": "workplaces",
				"parameters": [
					{
						"name": "doctor_id",
						"in": "query",
						"description": "",
						"required": true,
						"style": "form",
						"explode": true,
						"schema": {
							"type": "integer",
							"format": "int32",
							"example": 1
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
		"/doctor/workplace/staffList": {
			"get": {
				"tags": ["Doctor"],
				"summary": "staffList",
				"operationId": "staffList",
				"parameters": [
					{
						"name": "workplace_id",
						"in": "query",
						"description": "",
						"required": true,
						"style": "form",
						"explode": true,
						"schema": {
							"type": "integer",
							"format": "int32",
							"example": 1
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
...
		"/doctor/block/schedule": {
			"post": {
				"tags": ["Doctor"],
				"summary": "blockSchedule",
				"operationId": "blockSchedule",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/blockScheduleRequest"
							},
							"example": {
								"blocked_dates": [
									{
										"date": "2020-11-19",
										"day": "Monday"
									},
									{
										"date": "2020-11-20",
										"day": "Tuesday"
									}
								],
								"type": "schedule",
								"block_reason": "emergency"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false
			}
		},
		"/doctor/unBlock/schedule": {
			"post": {
				"tags": ["Doctor"],
				"summary": "unBlockSchedule",
				"operationId": "unBlockSchedule",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/unBlockScheduleRequest"
							},
							"example": {
								"dates": [
									{
										"date": "2020-11-19",
										"day": "Monday"
									},
									{
										"date": "2020-11-20",
										"day": "Tuesday"
									}
								],
								"type": "schedule"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false
			}
		},
		"/doctor/cancelAppointment": {
			"post": {
				"tags": ["Doctor"],
				"summary": "cancellAppointment",
				"operationId": "cancellAppointment",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/cancellAppointmentRequest"
							},
							"example": {
								"bookingID": 1,
								"reason": "emergency"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false
			}
		},
		"/doctor/addDelegates": {
			"post": {
				"tags": ["Doctor"],
				"summary": "AddDelegates",
				"operationId": "AddDelegates",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddDelegatesRequest"
							},
							"example": {
								"staff_id": 5,
								"workplaces_id": 9,
								"manageAppoinment": "1",
								"blockUnblockSchedue": 1,
								"changeSchedule": 1
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false
			}
		},
		"/doctor/removeDelegates": {
			"delete": {
				"tags": ["Doctor"],
				"summary": "Remove Delegates",
				"operationId": "RemoveDelegates",
				"parameters": [
					{
						"name": "staff_id",
						"in": "query",
						"description": "",
						"required": true,
						"style": "form",
						"explode": true,
						"schema": {
							"type": "integer",
							"format": "int32",
							"example": 5
						}
					},
					{
						"name": "workplaces_id",
						"in": "query",
						"description": "",
						"required": true,
						"style": "form",
						"explode": true,
						"schema": {
							"type": "integer",
							"format": "int32",
							"example": 9
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false
			}
		},
		"/doctor/drProfile": {
			"get": {
				"tags": ["Doctor"],
				"summary": "drProfile",
				"operationId": "drProfile",
				"parameters": [
					{
						"name": "doctor_id",
						"in": "query",
						"description": "",
						"required": true,
						"style": "form",
						"explode": true,
						"schema": {
							"type": "integer",
							"format": "int32",
							"example": 2
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
		"/doctor/updateSetupProfile": {
			"post": {
				"tags": ["Doctor"],
				"summary": "updateSetupProfile",
				"operationId": "PostupdateSetupProfile1",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/updateSetupProfileRequest2"
							},
							"example": {
								"professionalInformation": {
									"drInfo": {
										"birth_date": "1995-08-01",
										"gender": "male",
										"email": "dummy12345@gmail.com",
										"experience": 3
									},
									"specialites": [1]
								},
								"educationalQualification": {
									"mrdData": {
										"registration_number": "MRC201810084HRAC",
										"registration_year": 2018,
										"council_id": 1
									},
									"educationDetails": [
										{
											"qualification_id": 1,
											"university_id": 1,
											"year": 2017
										}
									]
								},
								"workplaceInformation": {
									"workplaces": [
										{
											"address": {
												"address_id": 2,
												"address": "Ram Tower, new Ambedkar Garden, Chembur - East , Mumbai, 400089",
												"location": {
													"latitude": 46.1052,
													"longitude": -46.383724
												},
												"locality": "sector 12",
												"city": "Mumbai",
												"pincode": "400341"
											},
											"workplace": {
												"workplace_id": 2,
												"workplace_name": "Ramo Hospptal",
												"time_per_appointment": 30,
												"consultation_fee": 500
											},
											"schedule": [
												{
													"day": "Monday",
													"start_time": "15:00",
													"end_time": "15:30"
												},
												{
													"day": "Monday",
													"start_time": "15:31",
													"end_time": "16:00"
												},
												{
													"day": "Monday",
													"start_time": "16:01",
													"end_time": "16:30"
												},
												{
													"day": "Wednesday",
													"start_time": "15:00",
													"end_time": "15:30"
												},
												{
													"day": "Wednesday",
													"start_time": "15:31",
													"end_time": "16:00"
												},
												{
													"day": "Wednesday",
													"start_time": "16:01",
													"end_time": "16:30"
												},
												{
													"day": "Friday",
													"start_time": "15:00",
													"end_time": "15:30"
												},
												{
													"day": "Friday",
													"start_time": "15:31",
													"end_time": "16:00"
												},
												{
													"day": "Friday",
													"start_time": "16:01",
													"end_time": "16:30"
												}
											]
										}
									]
								},
								"otherInformation": {
									"conventionDetails": {
										"medical_convention": "Allopathy",
										"prescription_limit": 14,
										"prescription_days_week_month": "Days"
									},
									"documentDetails": {
										"type": "ADHAAR",
										"number": "12FG12345678"
									},
									"features": [1, 2, 3]
								},
								"user_role": 2
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"updateSetupProfileRequest": {
				"title": "updateSetupProfileRequest",
				"required": ["laboratoryDetails", "user_role"],
				"type": "object",
				"properties": {
					"laboratoryDetails": {
						"$ref": "#/components/schemas/LaboratoryDetails"
					},
					"user_role": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"laboratoryDetails": {
						"workplaces": [
							{
								"address": {
									"address_id": 5,
									"address": "Rama Tower, new Gopal Garden, Chembur - East , Mumbai, 400089",
									"location": {
										"latitude": 46.1672,
										"longitude": -46.353724
									},
									"locality": "sector 11",
									"city": "Mumbai",
									"pincode": "400341"
								},
								"workplace": {
									"workplace_id": 1,
									"workplace_name": "Jara Plutonics Laboratory",
									"is_franchise": 1,
									"franchise_name": "Jara Industries",
									"phone_number": "9898167634",
									"license_number": "AC129ETBMQ56U",
									"gst_number": "12845NKYTLDS0",
									"delivery_customer": 1,
									"delivery_distance": 10
								},
								"employees": [
									{
										"first_name": "lab_assistant4",
										"middle_name": null,
										"last_name": "lab_lastname4",
										"email": "testlab4@gmail.com",
										"gender": "male",
										"birth_date": "1995-11-07",
										"contact_number": "+919066660652",
										"lab_or_pharma_employement_number": "1102",
										"document_type": "ADHAAR",
										"document_number": "12FG1222040",
										"edit_profile": 1
									}
								]
							}
						],
						"user_id": 8
					},
					"user_role": 4
				}
			},
...
			"Dashboard": {
				"title": "Dashboard",
				"required": ["success", "data", "message"],
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Data5"
						},
						"description": ""
					},
					"message": {
						"type": "string"
					}
				},
				"example": {
					"success": true,
					"data": [
						{
							"workplace_id": 1,
							"workplace_name": "Sunny11 Hospptal",
							"time_per_appointment": 20,
							"slots": [
								{
									"id": 1,
									"start_time": "10:00:00",
									"end_time": "12:00:00",
									"blocked_slots": []
								}
							],
							"patient_details": []
						},
						{
							"workplace_id": 2,
							"workplace_name": "AIIMS",
							"time_per_appointment": 30,
							"slots": [
								{
									"id": 3,
									"start_time": "15:00:00",
									"end_time": "17:00:00",
									"blocked_slots": []
								}
							],
							"patient_details": []
						}
					],
					"message": "Success"
				}
			},
...
			"updateDelegatesRequest": {
				"title": "updateDelegatesRequest",
				"required": [
					"staff_id",
					"workplaces_id",
					"manageAppoinment",
					"blockUnblockSchedue",
					"changeSchedule"
				],
				"type": "object",
				"properties": {
					"staff_id": {
						"type": "integer",
						"format": "int32"
					},
					"workplaces_id": {
						"type": "integer",
						"format": "int32"
					},
					"manageAppoinment": {
						"type": "integer",
						"format": "int32"
					},
					"blockUnblockSchedue": {
						"type": "integer",
						"format": "int32"
					},
					"changeSchedule": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"staff_id": 5,
					"workplaces_id": 9,
					"manageAppoinment": 1,
					"blockUnblockSchedue": 0,
					"changeSchedule": 0
				}
			},
			"addSetupProfileRequest2": {
				"title": "addSetupProfileRequest2",
				"required": [
					"professionalInformation",
					"educationalQualification",
					"workplaceInformation",
					"otherInformation",
					"user_role"
				],
				"type": "object",
				"properties": {
					"professionalInformation": {
						"$ref": "#/components/schemas/ProfessionalInformation1"
					},
					"educationalQualification": {
						"$ref": "#/components/schemas/EducationalQualification"
					},
					"workplaceInformation": {
						"$ref": "#/components/schemas/WorkplaceInformation1"
					},
					"otherInformation": {
						"$ref": "#/components/schemas/OtherInformation"
					},
					"user_role": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"professionalInformation": {
						"drInfo": {
							"birth_date": "1985-08-01",
							"gender": "male",
							"email": "dummydoc@gmail.com",
							"user": {
								"id": 3
							},
							"experience": 6
						},
						"specialites": [1]
					},
					"educationalQualification": {
						"mrdData": {
							"registration_number": "MRC200510084HRAC",
							"registration_year": 2005,
							"registration_council": 2
						},
						"educationDetails": [
							{
								"qualification_id": 1,
								"university_id": 1,
								"year": 2004
							}
						]
					},
					"workplaceInformation": {
						"workplaces": [
							{
								"address": {
									"address": "Sunny11 Tower, new Ambedkar Garden, Chembur - East , Mumbai, 400089",
									"location": {
										"latitude": 46.1052,
										"longitude": -46.383724
									},
									"locality": "sector 12",
									"city": "Mumbai",
									"pincode": "400341"
								},
								"workplace": {
									"workplace_name": "Sunny11 Hospptal",
									"time_per_appointment": 20,
									"consultation_fee": 200
								},
								"schedule": [
									{
										"day": "Monday",
										"start_time": "15:00",
										"end_time": "17:00"
									}
								]
							}
						]
					},
					"otherInformation": {
						"conventionDetails": {
							"medical_convention": "Allopathy",
							"prescription_limit": 14,
							"prescription_days_week_month": "Days"
						},
						"documentDetails": {
							"type": "ADHAAR",
							"number": "12FG12345678"
						},
						"features": [1, 2, 3]
					},
					"user_role": 2
				}
			},
...
		},
		"securitySchemes": {
			"ApiKeyAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"security": [
		{
			"ApiKeyAuth": []
		}
	]
}